Added possibility to skip tests dynamically by GitHub issue
